Create Authenticator app; update Projects
We currently have a React app deployed at
`https://tps.autark.xyz/tps-auth` which does basically what this
Authenticator does. Here's how it works:

1. Projects opens an popup which loads the GitHub OAuth login page
2. We have [a GitHub OAuth App][1], where we provide an Authorization
   Callback URL. This is configured to that `tps.autark.xyz` page.
3. This Authorizor receives a `code` from GitHub, which it posts back to
   Projects using `opener.postMessage`
4. Projects uses this `code` to hit `https://tps.autark.xyz/authenticate`,
   which holds onto the private key for our GitHub OAuth App and can
   thus load an API `token` for the user, so Projects can start making
   requests as this user.

This commit seeks to optimize three aspects of the current approach:

1. Code visibility/maintainability: rather than having a deployed app
   with seemingly no corresponding GitHub repository, this adds this key
   piece of our infrastructure directly to our main `open-enterprise`
   repo. It also simplifies this Authenticator, using a simple html page
   rather than a complex React app.
2. Design: rather than a sans-serif "loading" message, this gives a
   simple visual while the user is logged in, with a subtle technical
   status
3. Security: loading this app at a URL we own provides no additional
   security. Why? Because any app, not just Projects, could open a
   GitHub OAuth page and provide our CLIENT_ID. GitHub would then
   redirect back to our Authenticator, regardless of where it lives, and
   Authenticator would blindly pass the `code` it received back to
   `opener`. From there, a malicious app can easily POST that code to
   our `https://tps.autark.xyz/authenticate` endpoint and receive a
   token for the user. And then that malicious app could read a user's
   private repos and post comments as them, and it would look like these
   GitHub actions were mediated by us.

Unfortunately, the solution in this commit fails to address this
security concern. It tries to read `opener.location.pathname` so that it
can check the identity of the calling app, before passing `code` back to
it, but calling `opener.location.pathname` is forbidden, because this
Authenticator lives at a different domain than Projects. Or rather,
Projects is served from the same IPFS gateway, BUT since it is rendered
within an iframe, the browser treats it as living at a different URL
than the page in the popup.

IPFS hash of this version: QmSYYEAFJiPyPd6JGtreu8tQZwEjARDxGZpSnTEvex8K4C

  [1]: https://github.com/organizations/AutarkLabs/settings/applications/953918