Allocations/BudgetDetail: add Period Start & End
We weren't yet storing any period info in appState, so this also adds
logic to `store/events` to listen to `NewPeriod` events from our
contract and update a singular `period` key.

(As a reminder, yes, `period` is a global setting for all Budgets; it is
not specific to one. So keeping this appState key name singular is
exactly what we want.)

Someday we will support changing the period duration from the frontend,
and at that point we will need more logic in store/events to handle
that. But for now, this gets us through.